fix(check-in): 修正 check-in-detail 造成 vitest import-analysis 解析失敗
## Why is this necessary?

count-comments 測試套件因此無法執行（dev 也一樣紅）。根因不是
check-in-detail.tsx 本身的語法問題，而是 Next.js tsconfig 的
"jsx": "preserve" 讓 vitest（vite 8 / oxc）transform 保留 JSX 不轉譯，
測試經由 utils/count-comments.ts 間接 import 該 .tsx 時，
vite:import-analysis 便因原始 JSX 無法解析而整個 suite 掛掉。

## How does it address?

把純資料型別 ApiCommentNode 與 type guard isApiCommentNode 從
check-in-detail.tsx 抽出到新的 api-comment-node.ts（純 .ts 模組），
count-comments.ts 與其測試改從新模組 import，測試鏈路不再碰到任何
.tsx；check-in-detail.tsx 改為 re-export 同名符號，
persona/[id]/page.tsx、practices/[id]/page.tsx 等既有 import 路徑
與所有 runtime 行為完全不變。
